Smush => single Initializer, remove InitializeOptions
