+++ src/Microsoft.AspNetCore.Authentication.Cookies/CookieExtensions.cs
@@ -21,60 +21,8 @@ namespace Microsoft.Extensions.DependencyInjection
 
         public static IServiceCollection AddCookieAuthentication(this IServiceCollection services, string authenticationScheme, Action<CookieAuthenticationOptions> configureOptions)
         {
-            // Makes sure that DataProtectionProvider is always set.
-            services.TryAddEnumerable(ServiceDescriptor.Singleton<IInitializeOptions<CookieAuthenticationOptions>, EnsureDataProtection>());
-            services.Initialize<CookieAuthenticationOptions>(authenticationScheme, options =>
-            {
-                if (String.IsNullOrEmpty(options.CookieName))
-                {
-                    options.CookieName = CookieAuthenticationDefaults.CookiePrefix + authenticationScheme;
-                }
-                if (options.TicketDataFormat == null)
-                {
-                    if (options.DataProtectionProvider == null)
-                    {
-                        // This shouldn't happen normally due to the EnsureDataProtection initialize options.
-                        throw new InvalidOperationException("DataProtectionProvider must be provided.");
-                    }
-
-                    // Note: the purpose for the data protector must remain fixed for interop to work.
-                    var dataProtector = options.DataProtectionProvider.CreateProtector("Microsoft.AspNetCore.Authentication.Cookies.CookieAuthenticationMiddleware", authenticationScheme, "v2");
-                    options.TicketDataFormat = new TicketDataFormat(dataProtector);
-                }
-                if (options.CookieManager == null)
-                {
-                    options.CookieManager = new ChunkingCookieManager();
-                }
-                if (!options.LoginPath.HasValue)
-                {
-                    options.LoginPath = CookieAuthenticationDefaults.LoginPath;
-                }
-                if (!options.LogoutPath.HasValue)
-                {
-                    options.LogoutPath = CookieAuthenticationDefaults.LogoutPath;
-                }
-                if (!options.AccessDeniedPath.HasValue)
-                {
-                    options.AccessDeniedPath = CookieAuthenticationDefaults.AccessDeniedPath;
-                }
-            });
+            services.TryAddEnumerable(ServiceDescriptor.Singleton<IInitializeOptions<CookieAuthenticationOptions>, CookieAuthenticationInitializer>());
             return services.AddScheme<CookieAuthenticationOptions, CookieAuthenticationHandler>(authenticationScheme, configureOptions);
-        }
-
-        // Used to ensure that there's always a default data protection provider
-        private class EnsureDataProtection : IInitializeOptions<CookieAuthenticationOptions>
-        {
-            private readonly IDataProtectionProvider _dp;
-
-            public EnsureDataProtection(IDataProtectionProvider dataProtection)
-            {
-                _dp = dataProtection;
-            }
-
-            public void Initialize(string name, CookieAuthenticationOptions options)
-            {
-                options.DataProtectionProvider = options.DataProtectionProvider ?? _dp;
-            }
         }
     }
 }
