+++ src/Microsoft.AspNetCore.Authentication.OAuth/OAuthExtensions.cs
@@ -2,11 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Net.Http;
-using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.OAuth;
-using Microsoft.AspNetCore.DataProtection;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
 
@@ -23,48 +19,8 @@ namespace Microsoft.Extensions.DependencyInjection
             where TOptions : OAuthOptions, new()
             where THandler : OAuthHandler<TOptions>
         {
-            services.TryAddEnumerable(ServiceDescriptor.Singleton<IInitializeOptions<TOptions>, EnsureDataProtection<TOptions>>());
-            services.Initialize<TOptions>(authenticationScheme, options =>
-            {
-                if (options.Backchannel == null)
-                {
-                    options.Backchannel = new HttpClient(options.BackchannelHttpHandler ?? new HttpClientHandler());
-                    options.Backchannel.DefaultRequestHeaders.UserAgent.ParseAdd("Microsoft ASP.NET Core OAuth handler");
-                    options.Backchannel.Timeout = options.BackchannelTimeout;
-                    options.Backchannel.MaxResponseContentBufferSize = 1024 * 1024 * 10; // 10 MB
-                }
-
-                if (options.StateDataFormat == null)
-                {
-                    if (options.DataProtectionProvider == null)
-                    {
-                        // This shouldn't happen normally due to the EnsureDataProtection initialize options.
-                        throw new InvalidOperationException("DataProtectionProvider must be provided.");
-                    }
-
-                    var dataProtector = options.DataProtectionProvider.CreateProtector(
-                        typeof(THandler).FullName, authenticationScheme, "v1");
-                    options.StateDataFormat = new PropertiesDataFormat(dataProtector);
-                }
-            });
+            services.TryAddEnumerable(ServiceDescriptor.Singleton<IInitializeOptions<TOptions>, OAuthInitializer<TOptions, THandler>>());
             return services.AddRemoteScheme<TOptions, THandler>(authenticationScheme, authenticationScheme, configureOptions);
         }
-
-        // Used to ensure that there's always a default data protection provider
-        private class EnsureDataProtection<TOptions> : IInitializeOptions<TOptions> where TOptions : OAuthOptions
-        {
-            private readonly IDataProtectionProvider _dp;
-
-            public EnsureDataProtection(IDataProtectionProvider dataProtection)
-            {
-                _dp = dataProtection;
-            }
-
-            public void Initialize(string name, TOptions options)
-            {
-                options.DataProtectionProvider = options.DataProtectionProvider ?? _dp;
-            }
-        }
-
     }
 }
